+++ app/services/telegram/telegram_parser/parser/save_vacancy.py
@@ -1,6 +1,6 @@
 import datetime
-import uuid
 import logging
+import uuid
 
 from asgiref.sync import sync_to_async
