+++ app/services/utils/main.py
@@ -1,0 +1,7 @@
+def custom_title_filter_factory(filter_cls, title):
+    class Wrapper(filter_cls):
+        def __new__(cls, *args, **kwargs):
+            instance = filter_cls(*args, **kwargs)
+            instance.title = title
+            return instance
+    return Wrapper
